fix: BsonType not assignable to BsonArray
In MQL it is permissible to have a query like { countries: "Germany" }
where countries is a list of string in mongodb schema. We were
incorrectly marking such assignments as a possible flaw.

This commit fixes it by making sure that a given BsonType is assignable
to a BsonArray as well as long as the given BsonType is also assignable
to the underlying schema of the BsonArray