fixes nullable type assignability
This commit fixes the problem with a nullable type not being assigable
to the non-nullable similar type for example: BsonAnyOf(BsonString,
BsonNull) should be assignable to BsonString but it was not as we were
too strict with out type check. However this is needed because of how
Java driver boxes even the primitive types, making them nullable.

There is also a fix for lower precision values not being assignable to
higher precision values for example: BsonInt32 should be assignable to
BsonInt64 and BsonDouble should be assignable to BsonDecimal128 but it
was not.